# docs/mvp/API_SPECIFICATION_MVP.yaml
...
openapi: 3.0.3
info:
  title: Courtesy Inspection API - MVP
  description: |
    # Courtesy Inspection API Documentation - MVP Version
    
    Essential API specification for the Courtesy Inspection digital vehicle inspection platform.
    This MVP version focuses on core CRUD operations and basic workflow functionality.
    
    ## Authentication
    All API requests require authentication using JWT tokens obtained from the `/auth/login` endpoint.
    Include the token in the `Authorization` header: `Bearer <token>`
    
    ## Rate Limiting
    - Anonymous: 100 requests/hour
    - Authenticated: 1,000 requests/hour
    
    ## Pagination
    List endpoints support cursor-based pagination with `limit` and `cursor` parameters.
    
    ## Error Handling
    Errors follow RFC 7807 Problem Details format.
  version: 1.0.0-mvp
  contact:
    name: API Support
    email: api@courtesyinspection.com
servers:
  - url: https://api.courtesyinspection.com/v1
    description: Production server
  - url: https://staging-api.courtesyinspection.com/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Development server
...
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticate user and receive JWT tokens
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: mechanic@shop.com
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: SecurePassword123!
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: JWT access token
                  refresh_token:
                    type: string
                    description: JWT refresh token
                  expires_in:
                    type: integer
                    description: Token expiration time in seconds
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          description: Problem type identifier
        title:
          type: string
          description: Human-readable error title
        status:
          type: integer
          description: HTTP status code
        detail:
          type: string
          description: Human-readable error details
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        role:
          type: string
          enum: [super_admin, shop_manager, mechanic]
        shop_id:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
          enum: [active, inactive]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
security:
  - bearerAuth: []